requirements.txt
python-dotenv==1.0.0
aioconsole==0.3.1
ascii_magic==1.0.0
